Correção do presenter com lambda_header + controller corrigido
